Switch endpoint to integrations router + tests, too
