+++ src/xngin/apiserver/routers/admin_integrations/test_admin_integrations_api.py
@@ -8,21 +8,26 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from xngin.apiserver.conftest import expect_status_code
 from xngin.apiserver.routers.admin.admin_api_types import CreateOrganizationRequest
+from xngin.apiserver.routers.admin.test_admin_api import make_bandit_online_experiment
 from xngin.apiserver.routers.admin_integrations.admin_integrations_api_types import (
     SetConnectionToTurnRequest,
     SetTurnArmJourneyMappingRequest,
 )
 from xngin.apiserver.routers.common_api_types import (
+    Arm,
     ArmBandit,
     CreateExperimentRequest,
+    DesignSpecMetricRequest,
     LikelihoodTypes,
     MABExperimentSpec,
+    PreassignedFrequentistExperimentSpec,
     PriorTypes,
 )
 from xngin.apiserver.routers.common_enums import ExperimentsType
 from xngin.apiserver.sqla import tables
 from xngin.apiserver.testing.admin_api_client import AdminAPIClient
 from xngin.apiserver.testing.admin_integrations_api_client import AdminIntegrationsAPIClient
+from xngin.apiserver.testing.testing_dwh_def import TESTING_DWH_PARTICIPANT_DEF
 
 
 class FakeAsyncClient:
@@ -421,3 +426,80 @@ async def test_resetting_same_token_preserves_arm_journey_mapping(
 
     got = iaclient.get_turn_arm_journey_mapping(datasource_id=ds_id, experiment_id=experiment_id).data
     assert got.arm_to_journeys == mapping
+
+
+async def test_get_experiment_sample_calls_for_mab(
+    testing_datasource, aclient: AdminAPIClient, iaclient: AdminIntegrationsAPIClient
+):
+    """A MAB experiment's sample calls carry example calls with a type-correct outcome."""
+    experiment = await make_bandit_online_experiment(
+        aclient,
+        testing_datasource.datasource_id,
+        prior_type=PriorTypes.NORMAL,
+        reward_type=LikelihoodTypes.NORMAL,
+    )
+    sample_calls = iaclient.get_experiment_sample_calls(
+        datasource_id=testing_datasource.datasource_id,
+        experiment_id=experiment.config.experiment_id,
+    ).data
+    assert sample_calls is not None
+    assert [c.label for c in sample_calls.calls] == ["Get assignment", "Report outcome"]
+    outcome_call = sample_calls.calls[1]
+    assert outcome_call.method == "POST"
+    assert outcome_call.path.endswith("/outcome")
+    assert experiment.config.experiment_id in outcome_call.path
+    assert outcome_call.body == {"outcome": 1.5}  # NORMAL reward => real-valued example
+
+
+async def test_get_experiment_sample_calls_none_for_cmab(
+    testing_datasource, aclient: AdminAPIClient, iaclient: AdminIntegrationsAPIClient
+):
+    """CMAB assignment needs a context vector, so there are no sample calls yet (deferred)."""
+    experiment = await make_bandit_online_experiment(
+        aclient,
+        testing_datasource.datasource_id,
+        experiment_type=ExperimentsType.CMAB_ONLINE,
+        prior_type=PriorTypes.NORMAL,
+        reward_type=LikelihoodTypes.NORMAL,
+    )
+    sample_calls = iaclient.get_experiment_sample_calls(
+        datasource_id=testing_datasource.datasource_id,
+        experiment_id=experiment.config.experiment_id,
+    ).data
+    assert sample_calls is None
+
+
+async def test_get_experiment_sample_calls_frequentist_get_assignment_only(
+    testing_datasource, aclient: AdminAPIClient, iaclient: AdminIntegrationsAPIClient
+):
+    """Frequentist clients fetch assignments via the API but report outcomes through their DWH, so
+    they get a get-assignment example and no outcome call."""
+    datasource_id = testing_datasource.datasource_id
+    created = aclient.create_experiment(
+        datasource_id=datasource_id,
+        body=CreateExperimentRequest(
+            design_spec=PreassignedFrequentistExperimentSpec(
+                experiment_type=ExperimentsType.FREQ_PREASSIGNED,
+                experiment_name="test experiment",
+                description="test experiment",
+                table_name=TESTING_DWH_PARTICIPANT_DEF.table_name,
+                primary_key="id",
+                start_date=datetime(2024, 1, 1, tzinfo=UTC),
+                end_date=datetime.now(UTC) + timedelta(days=1),
+                arms=[Arm(arm_name="C", arm_description="C"), Arm(arm_name="T", arm_description="T")],
+                metrics=[DesignSpecMetricRequest(field_name="is_engaged", metric_pct_change=0.1)],
+                strata=[],
+                filters=[],
+                desired_n=10,
+            ),
+        ),
+        random_state=42,
+    ).data
+    aclient.commit_experiment(datasource_id=datasource_id, experiment_id=created.experiment_id)
+    sample_calls = iaclient.get_experiment_sample_calls(
+        datasource_id=datasource_id,
+        experiment_id=created.experiment_id,
+    ).data
+    assert sample_calls is not None
+    assert [c.label for c in sample_calls.calls] == ["Get assignment"]
+    assert all("outcome" not in c.path for c in sample_calls.calls)
