+++ src/xngin/apiserver/testing/admin_api_client.py
@@ -91,7 +91,6 @@ from xngin.apiserver.routers.common_api_types import (
     ListExperimentsResponse,
     PowerRequest,
     PowerResponse,
-    SampleCalls,
 )
 
 if TYPE_CHECKING:
@@ -3894,76 +3893,6 @@ class AdminAPIClient:
                 default_status=HTTPStatus.OK,
                 models={
                     HTTPStatus.OK: GetExperimentForUiResponse,
-                    HTTPStatus.BAD_REQUEST: HTTPExceptionError,
-                    HTTPStatus.UNAUTHORIZED: HTTPExceptionError,
-                    HTTPStatus.FORBIDDEN: HTTPExceptionError,
-                    HTTPStatus.NOT_FOUND: HTTPExceptionError,
-                    HTTPStatus.UNPROCESSABLE_CONTENT: AdminAPIClientHTTPValidationError,
-                },
-                path_params={
-                    "datasource_id": datasource_id,
-                    "experiment_id": experiment_id,
-                },
-                raise_if_not_default_status=raise_if_not_default_status,
-                client_exts=client_exts,
-            ),
-        )
-
-    @overload
-    def get_experiment_sample_calls(
-        self,
-        *,
-        datasource_id: str,
-        experiment_id: str,
-        raise_if_not_default_status: Literal[True] = True,
-        client_exts: AdminAPIClientExtensions | None = None,
-    ) -> AdminAPIClientResult[Literal[HTTPStatus.OK], SampleCalls | None]: ...
-    @overload
-    def get_experiment_sample_calls(
-        self,
-        *,
-        datasource_id: str,
-        experiment_id: str,
-        raise_if_not_default_status: Literal[False],
-        client_exts: AdminAPIClientExtensions | None = None,
-    ) -> (
-        AdminAPIClientResult[Literal[HTTPStatus.OK], SampleCalls | None]
-        | AdminAPIClientResult[Literal[HTTPStatus.BAD_REQUEST], HTTPExceptionError]
-        | AdminAPIClientResult[Literal[HTTPStatus.UNAUTHORIZED], HTTPExceptionError]
-        | AdminAPIClientResult[Literal[HTTPStatus.FORBIDDEN], HTTPExceptionError]
-        | AdminAPIClientResult[Literal[HTTPStatus.NOT_FOUND], HTTPExceptionError]
-        | AdminAPIClientResult[Literal[HTTPStatus.UNPROCESSABLE_CONTENT], AdminAPIClientHTTPValidationError]
-    ): ...
-    def get_experiment_sample_calls(
-        self,
-        *,
-        datasource_id: str,
-        experiment_id: str,
-        raise_if_not_default_status: bool = True,
-        client_exts: AdminAPIClientExtensions | None = None,
-    ) -> (
-        AdminAPIClientResult[Literal[HTTPStatus.OK], SampleCalls | None]
-        | AdminAPIClientResult[Literal[HTTPStatus.BAD_REQUEST], HTTPExceptionError]
-        | AdminAPIClientResult[Literal[HTTPStatus.UNAUTHORIZED], HTTPExceptionError]
-        | AdminAPIClientResult[Literal[HTTPStatus.FORBIDDEN], HTTPExceptionError]
-        | AdminAPIClientResult[Literal[HTTPStatus.NOT_FOUND], HTTPExceptionError]
-        | AdminAPIClientResult[Literal[HTTPStatus.UNPROCESSABLE_CONTENT], AdminAPIClientHTTPValidationError]
-    ):
-        return cast(
-            (
-                AdminAPIClientResult[Literal[HTTPStatus.OK], SampleCalls | None]
-                | AdminAPIClientResult[Literal[HTTPStatus.BAD_REQUEST], HTTPExceptionError]
-                | AdminAPIClientResult[Literal[HTTPStatus.UNAUTHORIZED], HTTPExceptionError]
-                | AdminAPIClientResult[Literal[HTTPStatus.FORBIDDEN], HTTPExceptionError]
-                | AdminAPIClientResult[Literal[HTTPStatus.NOT_FOUND], HTTPExceptionError]
-                | AdminAPIClientResult[Literal[HTTPStatus.UNPROCESSABLE_CONTENT], AdminAPIClientHTTPValidationError]
-            ),
-            self._route_handler(
-                path="/v1/m/datasources/{datasource_id}/experiments/{experiment_id}/sample-calls",
-                method=HTTPMethod.GET,
-                default_status=HTTPStatus.OK,
-                models={
-                    HTTPStatus.OK: SampleCalls | None,
                     HTTPStatus.BAD_REQUEST: HTTPExceptionError,
                     HTTPStatus.UNAUTHORIZED: HTTPExceptionError,
                     HTTPStatus.FORBIDDEN: HTTPExceptionError,
